Introduce the env tag option 'required'.
Fields with the `env` tag set can now also include an option string,
similar to the json package: e.g., `env:"VAR_NAME,required"`.
This commit adds support for the "required" option.

The option parsing logic is somewhat primitive.  Only a single option at
a time is respected, and only either no option or the `required` option.

Parse returns an error if the `required` option is set for the env tag and
the corresponding environment variable is empty.