Fixed handling of empty strings in the environment
Before, the default value was used for string variables that are set to
the empty string in the environment. Now, we use the value of the
environment, if and only if it is set, and the default value otherwise,
that is, if the variable is not set in the environment.