fix: blank env keys populating fields with undesirable values
Same issue as seen in caarlos0#152 with fields that don't specify an env key being populated by a blank key in the host environment.

This was resolved in 2ed4d36 but seems to have been re-introduced in e244b8a. Adding an additional case to the value check seems to solve it.